application/pyproject.toml
[tool.poetry]
name = "fc-sre-app"
version = "0.1.0"
description = ""
authors = ["Yuneui Jeong"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "~4.2"
djangorestframework = "^3.15"
gunicorn = "~22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
